Fixed test...
Fixed a typo in the internals visible to declaration.

Changed how the tests are accessing the redis cluster connection.

I think this looks a little nicer than it was before. It also sheds
light on how I initially made that stuff a little more complicated than
maybe it needed to be.